requirements.txt
streamlit
langchain
langchain-google-genai
langgraph
duckduckgo-search
python-dotenv